Solve race condition and remove value check
Having `queried` be incremented in the goroutines caused a race
condition where the value could be updated in several threads at the
same time losing updates. The easiest solution is to remove it from
the inside of query to put it in the response counting.

The check for response being bigger than 0 is useless as it will
always be either 0 or bigger, and adding 0 won't affect the results.