DVX-6950 Add sitemap package
This package is used to check sitemap files, which are compressed XML
files containing valid routes in a website. This package takes a list
of file URLs, fetches the files, uncompresses and parses them.

It reports the recency of file modification via timestamp and the
amount of entries present in the file as Prometheus metrics.