DVX-7220 Add package from cloudinary_exporter
This adds the `cloudinary` package from
[cloudinary_exporter](https://github.com/ifosch/cloudinary_exporter).
The cloudinary_exporter is the one currently used to get usage metrics from
cloudinary, but it is faulty due to how the Cloudinary's Admin API
behaves.
To fix that, this exporter is going to be replaced with a collector,
written as a dextre subcommand, which will send metrics to Push
Gateway.

Refs [DVX-7220](https://mydevex.atlassian.net/browse/DVX-7220)