Refactor reverseAddress to use strings.Builder
The strings.Builder introduced in Go 1.10 is the most efficient and
less error-prone way to do these operations and it allows to simplify
that code a bit.

In the process we remove a helper function with its tests and refactor
the remaining ones to use tables and subtests, and be able to run in
parallel.