Collect and expose Permission errors
We now collect and annotate the errors while adding or removing
Permissions to a Security Group. The method `Err()` allows to retrieve
the oldest error collected and removes it from the internal storage,
so we can iterate over all the errors that may have occurred in a long
series of operations.