pkg/dnsbl: Remove collector and cleanup package
Cleanup the interface of `pkg/dnsbl` removing useless helpers and
making `Query` the main interface. Also remove the Collector object as
it's outside the package responsibility and was deeply intertwined
with the rest of the package.

There is no proper way to test the standard libraries' DNS resoultion
without building a wrapper around the connections, so for now we just
drop those tests.

Adjust dependencies as a result of the removal.