pkg/dnsbl: Add simple wrapper around host lookup
In order to test the behaviour of `dnsbl.Query` we need to be able to
mock the DNS resolver. This is not possible with the standard library
in its current form.

This wrapper changes the lookup to use a function pointer instead of
calling the standard library methods directly, so we can override the
method implementation by providing a different function with same
interface. By default these function pointers call the standard
library methods, so behaviour should remain the same.

Reintroduce the tests we had before through the new functionality and
convert them to public interface tests.